agent-engine/requirements.txt
# Web framework
fastapi==0.133.0
uvicorn[standard]==0.34.0

# LangGraph / LangChain
langgraph>=1.0.2,<1.1.0
langchain>=1.2.0,<1.3.0
langchain-openai>=1.0.0,<2.0.0
langgraph-checkpoint-postgres>=2.0.0,<3.0.0

# Data validation
pydantic==2.10.6
python-decouple==3.8

# Database
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Observability
langfuse==3.14.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
